Conflicts:
	README.md
	stf/core/models.py

+++ stf/core/dataset.py
@@ -158,7 +158,7 @@ class Dataset(persistent.Persistent):
         binetflow_file_name = biargus_file_name_without_extension + '.binetflow'
         ra_path = Popen('bash -i -c "type ra"', shell=True, stdin=PIPE, stdout=PIPE).communicate()[0].split()[0]
         if ra_path:
-            (ra_data,ra_error) = Popen('ra -F ./confs/ra.conf -n -Z b -r '+biargus_file_name+' > '+binetflow_file_name, shell=True, stdin=PIPE, stdout=PIPE, stderr=PIPE).communicate()
+            (ra_data,ra_error) = Popen('ra -F ./confs/ra.conf -n -Z b -r '+biargus_file_name+'|sort -n > '+binetflow_file_name, shell=True, stdin=PIPE, stdout=PIPE, stderr=PIPE).communicate()
             if not ra_error:
                 # Add the new biargus file to the dataset
                 self.add_file(binetflow_file_name)
@@ -257,6 +257,10 @@ class Datasets(persistent.Persistent):
 
     def delete(self, dataset_id): 
         """ Delete a dataset from the list of datasets """
+        # Verify the decision
+        input = raw_input('Are you sure you want to delete this dataset? (YES/NO): ')
+        if input != 'YES':
+            return False
         # Before deleting the dataset, delete the connections
         from stf.core.connections import __group_of_group_of_connections__
         __group_of_group_of_connections__.delete_group_of_connections(dataset_id)
